// src/components/community/community.types.ts
export interface Settings {
  background_color: string;
  name: string;
  title: string;
  community_menu: '1' | '2' | '3' | '4';
  image_1: string | null;
  name_1: string;
  address_1: string;
  stars_1: any;
  review_1: string;
  name_2: string;
  image_2: string | null;
  address_2: string;
  stars_2: any;
  review_2: string;
  image_3: string | null;
  name_3: string;
  address_3: string;
  stars_3: any;
  review_3: string;
  image_4: string | null;
  name_4: string;
  address_4: string;
  stars_4: any;
  review_4: string;
}